+++ src/main/kotlin/com/alugames/models/Plan.kt
@@ -1,7 +1,9 @@
 package com.alugames.models
 
+import java.math.BigDecimal
+
 sealed class Plan(val type: String) {
-    open fun rentPrice(rent: GameRent): Double? {
-        return rent.game.price?.times(rent.periodRent.days)
+    open fun rentPrice(rent: GameRent): BigDecimal? {
+        return rent.game.price?.times(rent.periodRent.days.toBigDecimal())
     }
 }
